完成购物车功能开发

+++ shopping-service/shopping-api/src/main/java/com/gpmall/shopping/dto/AddCartRequest.java
@@ -9,8 +9,8 @@ import lombok.Data;
 @Data
 public class AddCartRequest extends AbstractRequest{
 
-    private Integer userId;
-    private Integer itemId;
+    private Long userId;
+    private Long itemId;
     private Integer num;
 
     @Override
